Haiku: Add 32/64 bit support
* Change handles on the java side to "long /*int*/".
* Use the respectively types on the native side.

+++ bundles/org.eclipse.swt/Eclipse SWT/haiku/org/eclipse/swt/graphics/Image.java
@@ -349,7 +349,7 @@ public Image(Device device, InputStream stream) {
 public Image(Device device, String filename) {
 	super(device);
 	if (filename == null) SWT.error(SWT.ERROR_NULL_ARGUMENT);
-	handle = (long /*int*/)HaikuImage.load(filename);
+	handle = HaikuImage.load(filename);
 	if (handle == 0) SWT.error(SWT.ERROR_NO_HANDLES);
 	type = SWT.BITMAP;
 	init();
